Ensure that there is enough capacity in table when reading data
By default, VectorPointTable has a capacity of 4096 items. When reading
files with more than 4096 points, this could lead to out of bounds write.

The problem is pdal::StreamPointTable::clear does not check that np < capacity()
and so allows writes to out of bounds region of memory.

To fix the issue, ensure that enough capacity to hold the input from the start.